Add files via upload
It specializes in producing text-based content, such as answering queries, delivering information, and creating tales with Groq Cloud.
Groq is a platform that allows you to run Rum Llama 3.1 on the cloud. The groq interference is extremely minimal since they employ LPU (Language Processing Unit), which means that when we ask queries, the responses come back quickly.